Fix handling of Types containing generic params
If a type like List<T> was used on a generic interface method the
Introspecter would not be able to tell that this type was the correct
one to implement the interface method and thus cause an error
when trying to use the method through the StaticInterface or
StaticDuckInterface classes. This has been fixed and types that
contain generic parameters are now handled correctly.