Fix handling of generic type constraints
If a generic type constraint forced a base-type that implemented at
least one interface, the generated generic type constraint would
include that interface, even though the interface was not an explicit
part of the generic type constraint. This resulted in a runtime
exception when calling a method with such a constraint through
DuckInterface, StaticDuckInterface or StaticInterface.